# backend/docs/swagger.yaml
basePath: /api
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@diktator.app
    name: Diktator Support
    url: https://github.com/starefossen/diktator
  description: A family-friendly spelling test application API
  license:
    name: MIT
    url: https://github.com/starefossen/diktator/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Diktator API
  version: "1.0"
paths:
...
  /api/wordsets/{id}/words/{word}/audio:
    get:
      consumes:
      - application/json
      description: Stream TTS audio for a specific word in a word set (generates on-demand,
        cached by browser)
      parameters:
      - description: Word Set ID
        in: path
        name: id
        required: true
        type: string
      - description: Word to generate audio for
        in: path
        name: word
        required: true
        type: string
      produces:
      - audio/ogg
      responses:
        "200":
          description: Audio file content (OGG Opus)
          schema:
            type: file
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Word set not found
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Failed to generate audio
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Stream Audio for Word
      tags:
      - wordsets
...
securityDefinitions:
  BearerAuth:
    description: 'OIDC JWT token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"